// dist/api/category/types.ts
/* eslint-disable */
// tslint:disable
/**
 * ECOMMERSE PROJECT MICROSERVICES
 * 
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import { Api } from '../../models'

export interface CategoryApi {
	getAllCategories: () => Promise<GetAllCategoriesResponse>
	postCategory: (request: Api.CategoryBody | undefined) => Promise<PostCategoryResponse>
	getCategory: (id: number) => Promise<GetCategoryResponse>
	updateCategory: (id: number, request: Api.CategoryBody | undefined) => Promise<UpdateCategoryResponse>
}

export type GetAllCategoriesResponse = GetAllCategories200Response | GetAllCategories500Response

export interface GetAllCategories200Response {
	status: 200
	body: string[]
	headers?: never
}

export interface GetAllCategories500Response {
	status: 500
	body: Api.GetAllCategories500Response
	headers?: never
}

export type PostCategoryResponse = PostCategory201Response | PostCategory404Response

export interface PostCategory201Response {
	status: 201
	body: Api.CategoryCreateResponse
	headers?: never
}

export interface PostCategory404Response {
	status: 404
	body: Api.CategoryCreationFailure
	headers?: never
}

export type GetCategoryResponse = GetCategory200Response | GetCategory500Response

export interface GetCategory200Response {
	status: 200
	body: string[]
	headers?: never
}

export interface GetCategory500Response {
	status: 500
	body: Api.GetCategory500Response
	headers?: never
}

export type UpdateCategoryResponse = UpdateCategory200Response | UpdateCategory500Response

export interface UpdateCategory200Response {
	status: 200
	body: Api.UpdateCategory200Response
	headers?: never
}

export interface UpdateCategory500Response {
	status: 500
	body: Api.UpdateCategory500Response
	headers?: never
}